+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project are documented in this file.
 
 ### Added
 
+- Added demo videos to docs
+- Added export data button logic
 - Added new email notification preferances
 - Implemented real pdf generation for certificates
 - Custom video player